[Issue-1626] Add detect os for get project id
