+++ packages/extension-base/src/services/request-service/types.ts
@@ -30,4 +30,3 @@ export interface AuthUrlInfo {
 
 export type AuthUrls = Record<string, AuthUrlInfo>;
 
-export type OSType = 'Mac OS' | 'iOS' | 'Windows' | 'Android' | 'Linux' | 'Unknown';
