+++ packages/extension-base/src/services/wallet-connect-service/constants.ts
@@ -1,14 +1,20 @@
 // Copyright 2019-2022 @subwallet/extension-base authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import { MobileOS } from '@subwallet/extension-base/background/KoniTypes';
 import { EIP155_SIGNING_METHODS, POLKADOT_SIGNING_METHODS, WalletConnectSigningMethod } from '@subwallet/extension-base/services/wallet-connect-service/types';
+import { getOS } from '@subwallet/extension-base/utils';
 import { SignClientTypes } from '@walletconnect/types';
 
-export const PROJECT_ID = '6da34c0b48164d27681924dd9a46d6be';
+export const PROJECT_ID_EXTENSION = '6da34c0b48164d27681924dd9a46d6be';
+export const PROJECT_ID_MOBILE = '6da34c0b48164d27681924dd9a46d6be';
 export const RELAY_URL = 'wss://relay.walletconnect.com';
+
+const isMobile = MobileOS.includes(getOS());
+
 export const DEFAULT_WALLET_CONNECT_OPTIONS: SignClientTypes.Options = {
   logger: 'debug',
-  projectId: PROJECT_ID,
+  projectId: !isMobile ? PROJECT_ID_EXTENSION : PROJECT_ID_MOBILE,
   relayUrl: RELAY_URL,
   metadata: {
     name: 'SubWallet',
