+++ packages/extension-base/src/utils/environment.ts
@@ -1,7 +1,7 @@
 // Copyright 2019-2022 @polkadot/extension authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { RuntimeEnvironment, RuntimeEnvironmentInfo } from '../background/KoniTypes';
+import { OSType, RuntimeEnvironment, RuntimeEnvironmentInfo } from '../background/KoniTypes';
 
 function detectRuntimeEnvironment (): RuntimeEnvironmentInfo {
   if (typeof window !== 'undefined' && typeof document !== 'undefined') {
@@ -59,3 +59,28 @@ function detectRuntimeEnvironment (): RuntimeEnvironmentInfo {
 }
 
 export const RuntimeInfo: RuntimeEnvironmentInfo = detectRuntimeEnvironment();
+
+export const getOS = (): OSType => {
+  const userAgent = window.navigator.userAgent;
+  // @ts-ignore
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access,@typescript-eslint/no-unsafe-assignment
+  const platform: string = window.navigator?.userAgentData?.platform || window.navigator.platform;
+  const macosPlatforms = ['Macintosh', 'MacIntel', 'MacPPC', 'Mac68K'];
+  const windowsPlatforms = ['Win32', 'Win64', 'Windows', 'WinCE'];
+  const iosPlatforms = ['iPhone', 'iPad', 'iPod'];
+  let os: OSType = 'Unknown';
+
+  if (macosPlatforms.indexOf(platform) !== -1) {
+    os = 'Mac OS';
+  } else if (iosPlatforms.indexOf(platform) !== -1) {
+    os = 'iOS';
+  } else if (windowsPlatforms.indexOf(platform) !== -1) {
+    os = 'Windows';
+  } else if (/Android/.test(userAgent)) {
+    os = 'Android';
+  } else if (/Linux/.test(platform)) {
+    os = 'Linux';
+  }
+
+  return os;
+};
